requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.8
cloudinary==1.28.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.2.9
django-bootstrap4==21.1
django-cloudinary-storage==0.3.0
django-crispy-forms==1.13.0
django-heroku==0.3.1
django-registration==3.1.2
djangorestframework==3.12.4
fontawesome==5.10.1.post1
gunicorn==20.1.0
idna==3.3
Pillow==8.4.0
psycopg2==2.8.6
psycopg2-binary==2.9.2
python-decouple==3.5
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
